moved extension event classes from rest to core (#2519)
moved responsibility of sending a success event to the extension service
