+++ bundles/core/org.eclipse.smarthome.core.extension.sample/META-INF/MANIFEST.MF
@@ -4,6 +4,7 @@ Bundle-Name: Eclipse SmartHome Sample Extension Service
 Bundle-SymbolicName: org.eclipse.smarthome.core.extension.sample
 Bundle-Version: 0.9.0.qualifier
 Import-Package: org.apache.commons.lang,
+ org.eclipse.smarthome.core.events,
  org.eclipse.smarthome.core.extension
 Bundle-Vendor: Eclipse.org/SmartHome
 Service-Component: OSGI-INF/*.xml
